Complete revamp of the tutorialsManager system
tutorialsManager v2. It is now way more powerful and easy to use, especially when wanting to create new tutorials that are not in direct link with others.
I wanted to do this new version for a while now, and since I wanted to add a tutorial frame for the new tab switch button, it was the occasion to do so.